Fix not being able to toggle analytics at runtime (#32)
* Bind analytics menu item to Analytics.enabled

The menu item's state should now always reflect the value of Analytics.enabled, and toggling it will also toggle Analytics.enabled and Crashes.enabled. We check that App Center is configured before accessing the value of Analytics.enabled, and initialise it if not. This is because MainMenu was checking the value for analyticsEnabled before App Center was being initialised in applicationDidFinishLaunching.

* Remove Preferences.swift

This code is no longer needed now that the analytics menu item is bound directly to the value of Analytics.enabled, which is already persisted to UserDefaults by App Center.